+++ src/types/CancelableComponent.ts
@@ -1,0 +1,6 @@
+type CancelableComponent<T> = {
+  value: T;
+  cancel: () => void;
+};
+
+export type { CancelableComponent };