+++ src/types/Paginated.ts
@@ -1,0 +1,8 @@
+type Paginated<T> = {
+  hasNextPage: boolean;
+  hasPrevPage: boolean;
+  totalPages: number;
+  data: T;
+};
+
+export type { Paginated };