+++ src/types/index.d.ts
@@ -1,3 +1,17 @@
 import type { Native } from '@/types/Native';
+import type { Paginated } from '@/types/Paginated';
+import type { CancelableComponent } from '@/types/CancelableComponent';
+import type { CRUDCompontent } from '@/types/CRUDCompontent';
+import type {
+  UseStateComponent,
+  DirectUseStateComponent,
+} from '@/types/UseStateComponent';
 
-export { Native };
+export {
+  Native,
+  Paginated,
+  CancelableComponent,
+  CRUDCompontent,
+  UseStateComponent,
+  DirectUseStateComponent,
+};